rename pruner preview modules to satisfy operator copy contract
